# webapp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OpenAPI definition",
    "description": "demobin API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/rest/admin-ui/colors": {
      "get": {
        "tags": [
          "color-resource"
        ],
        "operationId": "getList",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageColorDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "color-resource"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColorDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ColorDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "color-resource"
        ],
        "operationId": "patchMany",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "color-resource"
        ],
        "operationId": "deleteMany",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rest/admin-ui/colors/{id}": {
      "get": {
        "tags": [
          "color-resource"
        ],
        "operationId": "getOne",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ColorDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "color-resource"
        ],
        "operationId": "patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ColorDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "color-resource"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ColorDto"
                }
              }
            }
          }
        }
      }
    },
    "/rest/admin-ui/colors/by-ids": {
      "get": {
        "tags": [
          "color-resource"
        ],
        "operationId": "getMany",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ColorDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rest/admin-ui/productTypes": {
      "get": {
        "tags": [
          "product-type-resource"
        ],
        "operationId": "getList_1",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductTypeDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product-type-resource"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductTypeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProductTypeDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "product-type-resource"
        ],
        "operationId": "patchMany_1",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "product-type-resource"
        ],
        "operationId": "deleteMany_1",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/rest/admin-ui/productTypes/by-ids": {
      "get": {
        "tags": [
          "product-type-resource"
        ],
        "operationId": "getMany_1",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductTypeDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/rest/admin-ui/stocks": {
      "get": {
        "tags": [
          "stock-resource"
        ],
        "operationId": "getList_2",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageStockDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stock-resource"
        ],
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StockDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "stock-resource"
        ],
        "operationId": "patchMany_2",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonNode"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "stock-resource"
        ],
        "operationId": "deleteMany_2",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/rest/admin-ui/stocks/by-ids": {
      "get": {
        "tags": [
          "stock-resource"
        ],
        "operationId": "getMany_2",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/color/{id}": {
      "get": {
        "tags": [
          "color-controller"
        ],
        "operationId": "getColorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseEntity"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Sort": {
        "type": "object",
        "properties": {
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          }
        }
      },
      "ColorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PageColorDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "last": {
            "type": "boolean"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ColorDto"
            }
          },
          "first": {
            "type": "boolean"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "JsonNode": {
        "type": "object",
        "properties": {
          "integralNumber": {
            "type": "boolean"
          },
          "double": {
            "type": "boolean"
          },
          "valueNode": {
            "type": "boolean"
          },
          "floatingPointNumber": {
            "type": "boolean"
          },
          "bigInteger": {
            "type": "boolean"
          },
          "nodeType": {
            "type": "string",
            "enum": [
              "ARRAY",
              "BINARY",
              "BOOLEAN",
              "MISSING",
              "NULL",
              "NUMBER",
              "OBJECT",
              "POJO",
              "STRING"
            ]
          },
          "float": {
            "type": "boolean"
          },
          "int": {
            "type": "boolean"
          },
          "long": {
            "type": "boolean"
          },
          "textual": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          },
          "missingNode": {
            "type": "boolean"
          },
          "pojo": {
            "type": "boolean"
          },
          "number": {
            "type": "boolean"
          },
          "boolean": {
            "type": "boolean"
          },
          "null": {
            "type": "boolean"
          },
          "array": {
            "type": "boolean"
          },
          "binary": {
            "type": "boolean"
          },
          "containerNode": {
            "type": "boolean"
          },
          "short": {
            "type": "boolean"
          },
          "bigDecimal": {
            "type": "boolean"
          },
          "object": {
            "type": "boolean"
          }
        }
      },
      "ProductTypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "productDetailId": {
            "type": "integer",
            "format": "int64"
          },
          "productDetailWeight": {
            "type": "number",
            "format": "double"
          },
          "productDetailLength": {
            "type": "number",
            "format": "double"
          },
          "productDetailWidth": {
            "type": "number",
            "format": "double"
          },
          "productDetailHeight": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PageProductTypeDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "last": {
            "type": "boolean"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductTypeDto"
            }
          },
          "first": {
            "type": "boolean"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "StockDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "productTypeName": {
            "type": "string"
          },
          "colorName": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageStockDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "last": {
            "type": "boolean"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "sort": {
            "$ref": "#/components/schemas/Sort"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockDto"
            }
          },
          "first": {
            "type": "boolean"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "HttpMethod": {
        "type": "object",
        "properties": {}
      },
      "HttpRange": {
        "type": "object",
        "properties": {}
      },
      "Charset": {
        "type": "object",
        "properties": {
          "registered": {
            "type": "boolean"
          }
        }
      },
      "ContentDisposition": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "charset": {
            "$ref": "#/components/schemas/Charset"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "readDate": {
            "type": "string",
            "format": "date-time"
          },
          "attachment": {
            "type": "boolean"
          },
          "inline": {
            "type": "boolean"
          },
          "formData": {
            "type": "boolean"
          }
        }
      },
...
      "InetSocketAddress": {
        "type": "object",
        "properties": {
          "hostName": {
            "type": "string"
          },
          "unresolved": {
            "type": "boolean"
          },
          "address": {
            "$ref": "#/components/schemas/InetAddress"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "hostString": {
            "type": "string"
          }
        }
      },
      "MediaType": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "subtype": {
            "type": "string"
          },
          "parameters": {
            "type": "string"
          },
          "concrete": {
            "type": "boolean"
          },
          "charset": {
            "$ref": "#/components/schemas/Charset"
          },
          "qualityValue": {
            "type": "number",
            "format": "double"
          },
          "wildcardSubtype": {
            "type": "boolean"
          },
          "wildcardType": {
            "type": "boolean"
          },
          "subtypeSuffix": {
            "type": "string"
          }
        }
      },
      "Character": {
        "type": "object",
        "properties": {}
      },
      "Locale": {
        "type": "object",
        "properties": {
          "unicodeLocaleKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "country": {
            "type": "string"
          },
          "ISO3Language": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "displayVariant": {
            "type": "string"
          },
          "displayLanguage": {
            "type": "string"
          },
          "script": {
            "type": "string"
          },
          "unicodeLocaleAttributes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "displayCountry": {
            "type": "string"
          },
          "variant": {
            "type": "string"
          },
          "ISO3Country": {
            "type": "string"
          },
          "extensionKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Character"
            }
          },
          "displayScript": {
            "type": "string"
          }
        }
      },
      "LanguageRange": {
        "type": "object",
        "properties": {
          "range": {
            "type": "string"
          },
          "weight": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "HttpHeaders": {
        "type": "object",
        "properties": {
          "allow": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HttpMethod"
            }
          },
          "date": {
            "type": "integer",
            "format": "int64"
          },
          "expires": {
            "type": "integer",
            "format": "int64"
          },
          "accessControlMaxAge": {
            "type": "integer",
            "format": "int64"
          },
          "vary": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "origin": {
            "type": "string"
          },
          "ifUnmodifiedSince": {
            "type": "integer",
            "format": "int64"
          },
          "range": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HttpRange"
            }
          },
          "accessControlRequestMethod": {
            "$ref": "#/components/schemas/HttpMethod"
          },
          "ifModifiedSince": {
            "type": "integer",
            "format": "int64"
          },
          "accessControlAllowMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HttpMethod"
            }
          },
          "ifMatch": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "empty": {
            "type": "boolean"
          },
          "accessControlAllowHeaders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "accessControlAllowOrigin": {
            "type": "string"
          },
          "contentDisposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "host": {
            "$ref": "#/components/schemas/InetSocketAddress"
          },
          "connection": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "accessControlRequestHeaders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contentType": {
            "$ref": "#/components/schemas/MediaType"
          },
          "acceptLanguageAsLocales": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Locale"
            }
          },
          "upgrade": {
            "type": "string"
          },
          "acceptCharset": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Charset"
            }
          },
          "cacheControl": {
            "type": "string"
          },
          "ifNoneMatch": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pragma": {
            "type": "string"
          },
          "accept": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaType"
            }
          },
          "accessControlExposeHeaders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptLanguage": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LanguageRange"
            }
          },
          "ETag": {
            "type": "string"
          },
          "accessControlAllowCredentials": {
            "type": "boolean"
          },
          "contentLanguage": {
            "$ref": "#/components/schemas/Locale"
          },
          "contentLength": {
            "type": "integer",
            "format": "int64"
          },
          "location": {
            "type": "string",
            "format": "uri"
          },
          "lastModified": {
            "type": "integer",
            "format": "int64"
          },
          "acceptPatch": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaType"
            }
          }
        }
      },
      "HttpStatusCode": {
        "type": "object",
        "properties": {
          "3xxRedirection": {
            "type": "boolean"
          },
          "2xxSuccessful": {
            "type": "boolean"
          },
          "4xxClientError": {
            "type": "boolean"
          },
          "error": {
            "type": "boolean"
          },
          "1xxInformational": {
            "type": "boolean"
          },
          "5xxServerError": {
            "type": "boolean"
          }
        }
      },
      "ResponseEntity": {
        "type": "object",
        "properties": {
          "headers": {
            "$ref": "#/components/schemas/HttpHeaders"
          },
          "statusCodeValue": {
            "type": "integer",
            "format": "int32"
          },
          "statusCode": {
            "$ref": "#/components/schemas/HttpStatusCode"
          }
        }
      },
...
    }
  }
}